ecoproject/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
bleach==5.0.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
crypto==1.4.1
crypto101==2024.3.5.0
cssselect2==0.7.0
dateparser==1.2.0
distro==1.9.0
Django==3.2.14
django-bleach==3.1.0
django-cors-headers==4.4.0
django-qr-code==4.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
django-urlconf-export==1.1.1
djangorestframework==3.12.4
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.55.1
gobject==0.1.0
gTTS==2.5.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.7
IMAPClient==3.0.1
imgkit==1.2.3
importlib_metadata==8.4.0
inflect==7.3.1
iniconfig==2.0.0
jiter==0.9.0
more-itertools==10.4.0
mysqlclient==2.2.4
Naked==0.1.32
num2words==0.5.13
numpy==1.24.4
openai==1.76.0
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pip-install==1.3.5
pluggy==1.5.0
pprintpp==0.4.0
psycopg==3.2.1
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.15.0
pydantic==2.10.3
pydantic_core==2.27.1
pydyf==0.11.0
pyfortified-dateutil==0.1.1
PyPDF2==3.0.1
pyphen==0.16.0
pypng==0.20220715.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pytz-convert==0.2.9
PyYAML==6.0.2
qrcode==7.4.2
raisin.mysqldb==1.2
razorpay==1.4.2
regex==2024.7.24
reportlab==4.2.5
requests==2.32.3
safe-cast==0.3.4
segno==1.6.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.67.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
urlpattern==0.1.2
weasyprint==61.2
webencodings==0.5.1
whitenoise==6.7.0
XlsxWriter==3.2.0
xlwt==1.3.0
zipp==3.20.0
zopfli==0.2.3.post1
